Move to a single file approach until I can figure out why the server does not recognize modules
